When using the MongoDB backend, don't cleanup if result_expires is 0 or None. (#6462)
Fixes #6450.

+++ celery/backends/mongodb.py
@@ -248,6 +248,9 @@ class MongoBackend(BaseBackend):
 
     def cleanup(self):
         """Delete expired meta-data."""
+        if not self.expires:
+            return
+
         self.collection.delete_many(
             {'date_done': {'$lt': self.app.now() - self.expires_delta}},
         )
